requirements.txt
asgiref==3.6.0
Django==4.2
django-localflavor==4.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-decouple==3.8
sqlparse==0.4.4
